src/requirements.txt
absl-py==2.1.0
albumentations==1.2.1
astunparse==1.6.3
attrdict==2.0.1
attrs==23.2.0
boto3==1.34.42
botocore==1.34.42
bravado==11.0.3
bravado-core==6.1.1
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.48.1
future==0.18.3
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
idna==3.6
imageio==2.34.0
jenkspy==0.4.0
jmespath==1.0.1
joblib==1.3.2
jsonref==1.1.0
jsonschema==3.2.0
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
ml-dtypes==0.2.0
monotonic==1.6
msgpack==1.0.7
neptune-client==0.14.2
neptune-contrib==0.27.1
neptune-tensorflow-keras==0.9.9
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.2.5.6
nvidia-cuda-cupti-cu12==12.2.142
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu12==12.2.140
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cudnn-cu12==8.9.4.25
nvidia-cufft-cu12==11.0.8.103
nvidia-curand-cu12==10.3.3.141
nvidia-cusolver-cu12==11.5.2.141
nvidia-cusparse-cu12==12.1.2.141
nvidia-nccl-cu12==2.16.5
nvidia-nvjitlink-cu12==12.2.140
oauthlib==3.2.2
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.2
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.1
pyrsistent==0.20.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.0
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
swagger-spec-validator==3.0.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2024.2.12
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.0.7
websocket-client==1.7.0
Werkzeug==3.0.1
wrapt==1.14.1
